+++ examples/Axial.Refined.AotProbe/Program.fs
@@ -1,0 +1,9 @@
+open Axial.Refined
+
+[<EntryPoint>]
+let main _ =
+    Refine.positiveInt 42
+    |> Result.map PositiveInt.value
+    |> function
+        | Ok 42 -> 0
+        | other -> failwithf "Unexpected Refined probe result: %A" other